reimplement the ui_chatPromptColors cvar by creating global preprocessor defines instead of altering the color table
ui_chatPromptColors is a boolean cvar that tells whether to colorize the "Say to team:" and similar prompt strings with a respective color. this functionality was previously achieved the following way:
- menu files used the "^;" escape sequence to colorize the prompt strings, ie., they used "^;Say to team:".
- when a prompt menu was brought up, the value (which should have been a constant) of the 11th (corresponding to "^;") element of the g_color_table array was changed appropriately (based on the prompt variant and the ui_chatPromptColors setting).
- when the UI drew the prompt strings, the last setting of g_color_table[11] took effect for "^;".

other than the prompt strings, nothing in the menu files used "^;", but the use of "^;" was not safe due to it depending on a "garbage" value.

the reimplementation consists of the following:
- stop altering g_color_table[11].
- when loading menus, add the following global defines:
    ASSET_COLOR_STRING_SAY_TO_ALL "^2"
    ASSET_COLOR_STRING_SAY_TO_TEAM "^5"
    ASSET_COLOR_STRING_SAY_TO_ADMINS "^6"
    ASSET_COLOR_STRING_SAY_TO_IRC "^7"
    ASSET_COLOR_STRING_SAY_COMMAND "^7"
  if ui_chatPromptColors is true, and
    ASSET_COLOR_STRING_SAY_TO_ALL ""
    ASSET_COLOR_STRING_SAY_TO_TEAM ""
    ASSET_COLOR_STRING_SAY_TO_ADMINS ""
    ASSET_COLOR_STRING_SAY_TO_IRC ""
    ASSET_COLOR_STRING_SAY_COMMAND ""
  otherwise.
- make the ui_chatPromptColors cvar latched.
- in the menu files, use the ASSET_COLOR_STRING_SAY_* defines instead of "^;".

unfortunately, from now on, the UI needs to be reloaded for the changes to the ui_chatPromptColors cvar to take effect, and also, both of the "Say:" and "^7Say:" strings need to be individually translated.