Remove USE_ALPHA_TEST shader macro.
Instead pass the alpha threshold as a uniform float. To allow both
testing for greater than and less than threshold, a little math
trick using the abs is used, but abs is a free operation on GPUs,
so it should be just as fast as a simple comparison.

Testing for greater or equal / less than or equal is not possible,
but it hardly makes a difference on interpolated floats.