Menu color change, some clean up of the .menu files, menu order change
Changelog:
Removed the redundant use of "forecolor 1 1 1 1" from all menu files.
Changed the UI color scheme to closer match that of the website, logo, and game background.
Rearranged the options menu so that the graphical settings make more sense.
Created a new system sub category called "Effects"
Updated the presets to include all graphical effects and insured that all pre-exsisting effects had 'uiScript glCustom' so that the preset is correctly changed to Custom upon changing a setting.
Added in motion blur as a setting