scale/rotate/offset models at prepare time
- build iqm from iqe at prepare time and scale/rotate/offset them
- adjust model.cfg files since engine does not have to scale/rotate them anymore
- gitignore paktrace and iqm files

since the model are now rotated, scaled and offseted at prepare time,
map editors like netradiant can display them nicely.

note that the medistation is using a different orientation when
animated or not, so the medistation is rotated the way it displays
correctly when not animated (for tools like netradiant) and the
engine rotates it since it renders it animated.