Rewrite circle menus using Lua
The circle is now centered at the center of the screen (where the cursor
is placed).

A Lua script is used to generate the layout instead of a C++ class.

Another enhancement: in most cases, if an option becomes available while
the menu is open (e.g. because the team got more momentum), you will now
be able to select in in the circlemenu (even though it still won't be
displayed as available).